+++ packages/form-elements/CHANGELOG.md
@@ -1,7 +1,11 @@
 # Changelog
 
-## 0.1.0
-- Initial release
+# 0.2.0
+- Migrate tag names from `dbp-*-element` to `dbp-form-*-element`, for example
+  `dbp-string-element` to `dbp-form-string-element`
 
 # 0.1.1
 - Version bump for conformity
+
+## 0.1.0
+- Initial release